move tfstate.go into plugin.go
