+++ plugin.go
@@ -1,10 +1,15 @@
 package ecspresso
 
-import "fmt"
+import (
+	"errors"
+	"fmt"
+
+	"github.com/kayac/go-config/tfstate"
+)
 
 type ConfigPlugin struct {
-	Name   string `yaml:"name"`
-	Config map[string]interface{}
+	Name   string                 `yaml:"name"`
+	Config map[string]interface{} `yaml:"config"`
 }
 
 func (p ConfigPlugin) Setup(c *Config) error {
@@ -14,5 +19,18 @@ func (p ConfigPlugin) Setup(c *Config) error {
 	default:
 		return fmt.Errorf("plugin %s is not available", p.Name)
 	}
+	return nil
+}
+
+func setupPluginTFState(p ConfigPlugin, c *Config) error {
+	path, ok := p.Config["path"].(string)
+	if !ok {
+		return errors.New("tfstate plugin requires path for tfstate file as string")
+	}
+	funcs, err := tfstate.Load(path)
+	if err != nil {
+		return err
+	}
+	c.templateFuncs = append(c.templateFuncs, funcs)
 	return nil
 }
