+++ tests/terraform.tfstate
@@ -1,5 +1,5 @@
 {
-    "version": 1,
+    "version": 4,
     "terraform_version": "0.12.23",
     "serial": 1,
     "lineage": "054d7292-3d84-0584-4590-24d6f3b17399",
